Conflicts:
	composer.lock
	src/KAC/SiteBundle/Resources/config/services.yml
	src/KAC/SiteBundle/Resources/public/css/basket/basket.scss
	src/KAC/SiteBundle/Resources/views/Basket/summary.html.twig

+++ src/KAC/SiteBundle/Entity/Image.php
@@ -493,7 +493,10 @@ class Image
     {
         foreach ($this->filesForRemoval as $fileForRemoval)
         {
-            unlink($fileForRemoval);
+            if(file_exists($fileForRemoval))
+            {
+                unlink($fileForRemoval);
+            }
         }
     }
